+++ SudokuSolver/Solvers/BacktrackSolvers/CellPosition.cs
@@ -2,11 +2,11 @@
 {
     public class CellPosition
     {
-        public int X { get; set; }
-        public int Y { get; set; }
+        public byte X { get; set; }
+        public byte Y { get; set; }
         public int Possibilities { get; set; }
 
-        public CellPosition(int x, int y, int possibilities)
+        public CellPosition(byte x, byte y, int possibilities)
         {
             X = x;
             Y = y;
